TP1/requirements.txt
azure-mgmt-resource>=18.0.0
azure-mgmt-storage
azure-identity>=1.5.0
python-dotenv
azure-storage-file-datalake